refactor: split firestore god repository to separate repositories
