+++ app/src/main/java/com/pv239/beelocal/data/repository/SocialRepository.kt
@@ -2,7 +2,6 @@ package com.pv239.beelocal.data.repository
 
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.firestore.DocumentSnapshot
-import com.pv239.beelocal.domain.FirestoreRepository
 import com.pv239.beelocal.domain.Page
 import com.pv239.beelocal.model.FeedEntry
 import com.pv239.beelocal.model.User
@@ -11,12 +10,16 @@ import javax.inject.Singleton
 
 /**
  * Focused repository for the Social screen.
- * Delegates Firestore calls to [FirestoreRepository] and exposes
- * higher-level use-case methods consumed by [SocialViewModel].
+ *
+ * Composes the smaller, per-domain repositories ([UserRepository],
+ * [FollowRepository], [FeedRepository]) into the few use-cases the social UI
+ * actually needs, and consistently scopes them to the signed-in user.
  */
 @Singleton
 class SocialRepository @Inject constructor(
-    private val firestoreRepository: FirestoreRepository,
+    private val userRepository: UserRepository,
+    private val followRepository: FollowRepository,
+    private val feedRepository: FeedRepository,
     private val auth: FirebaseAuth,
 ) {
     /**
@@ -33,7 +36,7 @@ class SocialRepository @Inject constructor(
     // Current user
     // -------------------------------------------------------------------------
 
-    suspend fun getCurrentUser(): User? = firestoreRepository.getUser(currentUserId)
+    suspend fun getCurrentUser(): User? = userRepository.getUser(currentUserId)
 
     // -------------------------------------------------------------------------
     // Search
@@ -49,7 +52,7 @@ class SocialRepository @Inject constructor(
     ): Page<User> {
         if (query.isBlank()) return Page(emptyList(), null, hasMore = false)
         val uid = currentUserId
-        val page = firestoreRepository.searchUsers(query.trim(), lastVisible)
+        val page = userRepository.searchUsers(query.trim(), lastVisible)
         // Filter out self
         return page.copy(items = page.items.filter { it.id != uid })
     }
@@ -62,7 +65,7 @@ class SocialRepository @Inject constructor(
     suspend fun getFriends(): List<User> {
         val user = getCurrentUser() ?: return emptyList()
         return user.friends.mapNotNull { friendId ->
-            firestoreRepository.getUser(friendId)
+            userRepository.getUser(friendId)
         }
     }
 
@@ -79,11 +82,11 @@ class SocialRepository @Inject constructor(
     suspend fun requestFollow(targetUserId: String): Boolean {
         val me = getCurrentUser()
             ?: throw IllegalStateException("Current user document not found")
-        return firestoreRepository.requestFollow(fromUser = me, toUserId = targetUserId)
+        return followRepository.requestFollow(fromUser = me, toUserId = targetUserId)
     }
 
     suspend fun removeFriend(friendId: String) {
-        firestoreRepository.removeFriend(currentUserId = currentUserId, friendId = friendId)
+        userRepository.removeFriend(currentUserId = currentUserId, friendId = friendId)
     }
 
     // -------------------------------------------------------------------------
@@ -97,6 +100,6 @@ class SocialRepository @Inject constructor(
     suspend fun getFriendsFeed(): Page<FeedEntry> {
         val user = getCurrentUser()
             ?: return Page(emptyList(), null, hasMore = false)
-        return firestoreRepository.getFriendsFeed(user.friends)
+        return feedRepository.getFriendsFeed(user.friends)
     }
 }
